fix: clear the warnings that pointed at real reconstruction defects
The green build left 881 warnings. Seventy-eight of them were the
classes that indicate a wrong reconstruction rather than period-accurate
code, and each was settled against the binary rather than silenced.
Four turned out to be live bugs.

RBTermView assigned five freshly-created subviews straight into assign
properties, so ARC released each one on the next line and left the ivar
dangling. The binary builds them the other way round — configure a
local, addSubview:, then store — and its setters neither retain nor
release. Restored that order and made the five properties weak.
RBApplilinkView's indicator is the opposite case: the binary stores the
ivar first and configures through it, so the +1 has to live in the
field and the property is strong.

RBCampaignViewController wrapped an NSURL in URLWithString:, which
takes a string, so every campaign link resolved to nil and silently
failed to open. The binary passes linkURL straight to openURL:.

The unlock package description formats a 32-bit ivar with %zd. Both
halves are faithful — the format literal really is %zd and the getter
really is ldr w0 — but they only agreed while ssize_t was 32 bits.
Widened the arguments at the call, the same 32-to-64 hazard as the
audio source-record truncation.

AppDelegate declared setServerData:andB: as returning void while the
implementation and the binary both return BOOL, hiding the result from
every caller.

The rest are annotations and types brought into line with the binary:
twenty-two framework initialiser overrides that cannot widen
nullability, five MusicData name properties that its MusicDataFromDoc
subclass overrides with a bare return nil, locals retyped to what their
producers actually return, three duplicate theme enumerations folded
into RBUserSettingDataTheme, and a switch given the two enumerators the
binary handles by doing nothing.

Where the binary's own behaviour is what trips the warning it is kept
and commented: three pasteboard calls really do pass nil to a parameter
Apple marks nonnull, three format strings really are runtime strings
with no arguments, and one parsed value really is computed and
discarded.

+++ Project/GameSystem/src/note_replay.mm
@@ -28,7 +28,7 @@ constexpr int kNoteKindTap = 0;
 // the live note list: matching-side notes take the recorded judge, JR flag, long rate, and (for
 // slide notes) each sub-point judge; a skipped just-reflec entry back-fills an unmatched tap head as
 // a just-reflec miss.
-void ApplyReplayForSide(ReplayData *pReplay, unsigned int nGhostSide) {
+void ApplyReplayForSide(ReplayData *pReplay, int nGhostSide) {
     NoteEffectMgr *pMgr = NoteEffectMgr::shared();
     int nLiveIndex = 0;
     int nPendingJustReflec = 0;
@@ -111,7 +111,8 @@ void ApplyReplayGhostToNotes() {
     // opposite side.
     const int nPlayColor = pGameSystem->GetPlayColor();
     const int nFirstSide = NoteEffectMgr::shared()->FindNoteByIndex(0)->GetSide();
-    const unsigned int nGhostSide =
-        nFirstSide == nPlayColor ? nPlayColor : static_cast<unsigned int>(nPlayColor == 0 ? 1 : 0);
+    // Every side value in play (the play colour, a note's side, the no-side sentinel 3) is a
+    // non-negative int, so the ghost side is one too.
+    const int nGhostSide = nFirstSide == nPlayColor ? nPlayColor : (nPlayColor == 0 ? 1 : 0);
     ApplyReplayForSide(pReplay, nGhostSide);
 }
